when the minimum detail is set to max in terrain, which outputs actual number to vertices to accomodate the heightmap.
at 1024 i got 238 fps, and with 0 change in fps when editing terrain.

the higher settings causes a lot of increase in triangles/vertices which causes a lot of CPU overhead which results in bad perfomance.

the 4096, produces 31 million triangles, while the 1024 produces only 2 million triangles.
thats a huge difference.

I will be adding tweening so the your current system of placing objects will need a little update in short to use terrain tools api to get the height of the terrain at the specific point instead of raycasting.
which will allow to update obtain the height nearly realtime ,
and match the tweeening.